requirements.txt
click==8.1.3
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.1
MySQL-connector-python==8.0.29
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.1
Flask-Cors==3.0.10
redis==4.3.4
bcrypt==3.2.2
